+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Version 2.2  (2018-02-xx)
 - Issue #414: Add additional `Network.Host` methods.
 - Issue #415: Added `ProcessContext` static class to determine the context of the current process.
 - Issue #422: Add `Copy-FileWithProgress.ps1` demonstrating file copy with progress report.
+- Issue #423: Add `Copy-DirectoryWithProgress.ps1` demonstrating directory copy with progress report.
 
 ### Improvements